Create README.md
Keyword Driven Framework Using Selenium WebDriver

Online Shopping site 
Domain - Iceland maintance and checking for functionality 
Let’s create a simple design of a Keyword Driven Framework with the help of a scenario. The scenario will automate the end to end flow of a web application.

Scenario to be automated:

1. Open a Browser
2. Navigate to URL https://www.iceland.co.uk/
3. Click Signin link
3. Enter username - inro4ebony@gmail.com
4. Enter password - @45adminD
5. Click on the Signin  button
6. Click on the Signout button
7. Close the browser

Step1: To design a Keyword Driven Framework, first identify all the actions to be performed for automated testing of an application. In the above scenario, there are seven actions that have to be performed.

Step 2: Once all actions are identified, the next step is to create a keyword map table in the Excel sheet. Keyword map table is a table in an Excel sheet that defines all the keywords available for test automation projects.

The table below shows the keyword map table for the above scenarioKeyword Driven Testing Framework